services/agent_orchestrator/requirements.txt
requests==2.32.3
pydantic==2.9.2
python-dotenv==1.0.1
boto3==1.35.36
GitPython==3.1.43
